Combined all product mapping extensions into a single "ProductMappingExtensions" class and made it internal
