+++ src/Merchello.Web/Merchello.Web.csproj
@@ -224,11 +224,7 @@
   <ItemGroup>
     <Compile Include="ApplicationContextExtensions.cs" />
     <Compile Include="Cache\HttpCachingBacker.cs" />
-    <Compile Include="Models\ContentEditing\IProductAttributeExtensions.cs" />
-    <Compile Include="Models\ContentEditing\IProductOptionExtensions.cs" />
-    <Compile Include="Models\ContentEditing\IProductVariantExtensions.cs" />
-    <Compile Include="Models\ContentEditing\IProductExtensions.cs" />
-    <Compile Include="Models\ContentEditing\ProductVariantDisplayExtensions.cs" />
+    <Compile Include="Models\ContentEditing\ProductMappingExtensions.cs" />
     <Compile Include="ProductQuery.cs" />
     <Compile Include="CustomerExtensions.cs" />
     <Compile Include="Editors\CustomerApiController.cs" />
@@ -239,13 +235,10 @@
     <Compile Include="CustomerContext.cs" />
     <Compile Include="Models\Basket.cs" />
     <Compile Include="Models\ContentEditing\ExamineDisplayExtensions.cs" />
-    <Compile Include="Models\ContentEditing\ProductAttributeDisplayExtensions.cs" />
     <Compile Include="Models\ContentEditing\ProductAttributeDisplay.cs" />
     <Compile Include="Models\ContentEditing\ProductDisplay.cs" />
     <Compile Include="Models\ContentEditing\ProductDisplayBase.cs" />
-    <Compile Include="Models\ContentEditing\ProductDisplayExtensions.cs" />
     <Compile Include="Models\ContentEditing\ProductOptionDisplay.cs" />
-    <Compile Include="Models\ContentEditing\ProductOptionDisplayExtensions.cs" />
     <Compile Include="Models\ContentEditing\ProductVariantDisplay.cs" />
     <Compile Include="Models\IBasket.cs" />
     <Compile Include="Mvc\MerchelloTemplatePage.cs" />
